websocket: expose WebSocketWrapped for only wrap websocket
